refactor: modularize procCTRL TUI rendering and logic
- Extracted TUI rendering logic into `procCTRL_TUI_render.c` with modular row-based rendering helpers.
- Extracted keyboard input handling into `procCTRL_TUI_input.c`.
- Created `procCTRL_TUI_internal.h` to define `procctrl_context_t` and share state between components.
- Refactored `procCTRL_TUI.c` to use new modular init/cleanup helpers, reducing main function LOC and complexity.
